requirements/dev_requirements.txt
pytest-runner
pytest-pylint
pytest-pycodestyle
pytest-pep257
pytest-cov
pytest-django
pytest
pylint
pycodestyle
pep257
codecov